Don't stroke a path if it's not specified in the SVG
The code in the svg renderer is using a QPen with it's brush
set to Qt::NoBrush as a hack to indicate that we don't have a
stroke to paint, but to nevertheless use QPen to keep track of
state. So when rendering, we have to check for the pen having
no brush and in that case not draw anything.

Task-number: QTBUG-36767
Change-Id: Ib622c7d0108e7360437f57853f11c65190c16773
Reviewed-by: Laszlo Agocs <laszlo.agocs@digia.com>